refactor sds component
